requirements.txt
bs4
readability-lxml
lxml_html_clean